Corrected a stubbed modules problem for RN fiber bundle (#9784)
